Elastex v0.1.2 Elastex

This module contains the run functions used to make a HTTP request

Functions

build(req, conn)

Specs

build(%Elastex.Builder{action: term, body: term, headers: term, id: term, index: term, method: term, options: term, params: term, type: term, url: term}, map) :: %Elastex.Builder{action: term, body: term, headers: term, id: term, index: term, method: term, options: term, params: term, type: term, url: term}

Builds request for http call.

run(req)

Specs

run(%Elastex.Builder{action: term, body: term, headers: term, id: term, index: term, method: term, options: term, params: term, type: term, url: term}) :: any

Runs the request using the elastex url from config file

run(req, conn)

Specs

run(%Elastex.Builder{action: term, body: term, headers: term, id: term, index: term, method: term, options: term, params: term, type: term, url: term}, map) :: any

Runs the request using the url from the conn map parameter