Elastex v0.1.2 Elastex.Cluster

Functions

health()

Gets a very simple status on the health of a cluster.

health(index)

node_hot_threads(node)

Retrieves current hot threads on each node in the cluster.

node_info()

Retrieve one or more (or all) of the cluster nodes statistics.

node_info(attributes)

node_info(attributes, nodes)

node_stats()

Retrieve one or more (or all) of the cluster nodes statistics.

node_stats(stats)

node_stats(stats, nodes)

pending_tasks()

Retrieves pending tasks from a cluster wide perspective.

reroute(body)

Executes a reroute allocation command including specific commands.

state()

Gets comprehensive state information of cluster.

state(index, metrics)

stats()

Retrieves statistics from a cluster wide perspective.