Elastex v0.1.2 Elastex.Search

Follows the Elasticsearch Search API.

Types

body :: map | struct | nil

Functions

count(index, type)

Executes a query to get the number of matches for that query.

count(body, index, type)

explain(body, index, type, id)

Computes a score explanation for a query and a specific document.

multi_search(query_builders)

Specs

multi_search([Elastex.Builder.t]) :: Elastex.Builder.t

Executes several search request in a single call.

Examples

iex> query_builders = [
...> Elastex.Search.query(%{hello: "world"}),
...> Elastex.Search.query(%{hello: "world"}, "twitter", "tweet")
...> ]
iex> Elastex.Search.multi_search(query_builders)
%Elastex.Builder {
  url: "_msearch",
  body: "{}\n{\"hello\":\"world\"}\n{\"index\":\"twitter\"}\n{\"hello\":\"world\"}\n",
  method: :post,
  action: :multi_search
}

params(builder, params)

Specs

params(Elastex.Builder.t, Keyword.t) :: Elastex.Builder.t

query()

Specs

query :: Elastex.Builder.t

Executes a search query.

Useful for URI searches.

Examples

iex> Elastex.Search.query
%Elastex.Builder {
  url: "_search",
  method: :post,
  action: :search_query,
}

with params

iex> Elastex.Search.query |> Elastex.Search.params([q: "user:kimchy"])
%Elastex.Builder {
  url: "_search",
  method: :post,
  action: :search_query,
  params: [q: "user:kimchy"]
}

query(body)

Specs

query(body) :: Elastex.Builder.t

Executes a search query with a body.

Examples

iex> body = %{query: %{term: %{user: "kimchy"}}}
iex> Elastex.Search.query(body)
%Elastex.Builder {
  url: "_search",
  body: %{query: %{term: %{user: "kimchy"}}},
  method: :post,
  action: :search_query,
}

query(body, index)

Specs

query(body, String.t) :: Elastex.Builder.t

Executes a search query with a body and index.

Examples

iex> body = %{query: %{term: %{user: "kimchy"}}}
iex> Elastex.Search.query(body, "twitter")
%Elastex.Builder {
  url: "twitter/_search",
  body: %{query: %{term: %{user: "kimchy"}}},
  method: :post,
  index: "twitter",
  action: :search_query
}

query(body, index, type)

Specs

query(body, String.t, String.t) :: Elastex.Builder.t

Executes a search query with a body, index, and type.

Examples

iex> body = %{query: %{term: %{user: "kimchy"}}}
iex> Elastex.Search.query(body, "twitter", "tweet")
%Elastex.Builder {
  url: "twitter/tweet/_search",
  body: %{query: %{term: %{user: "kimchy"}}},
  index: "twitter",
  type: "tweet",
  action: :search_query,
  method: :post
}

shards(index)

Specs

shards(String.t) :: Elastex.Builder.t

Returns indices and shards that a search request would be executed against using an index.

Examples

iex> Elastex.Search.shards("twitter")
%Elastex.Builder {
  url: "twitter/_search_shards",
  method: :get,
  index: "twitter",
  type: ""
}

shards(index, type)

Specs

shards(String.t, String.t) :: Elastex.Builder.t

Returns indices and shards that a search request would be executed against using an index and a type.

Examples

iex> Elastex.Search.shards("twitter", "tweet")
%Elastex.Builder {
  url: "twitter/tweet/_search_shards",
  method: :get,
  index: "twitter",
  type: "tweet"
}

suggest(body)

Specs

suggest(body) :: Elastex.Builder.t

Suggests similar looking terms based on a provided text.

Examples

iex> body = %{"my-suggestion" => %{text: "hi", term: %{field: "body"}}}
iex> Elastex.Search.suggest(body)
%Elastex.Builder {
  url: "_suggest",
  body: %{"my-suggestion" => %{text: "hi", term: %{field: "body"}}},
  method: :post
}

template(body)

Specs

template(body) :: Elastex.Builder.t

Allows use of mustache templating language to pre-render search requests.

Examples

iex> body = %{inline: %{query: %{match: %{title: "{{query_string}}"}}}}
iex> Elastex.Search.template(body)
%Elastex.Builder {
  url: "template",
  body: %{inline: %{query: %{match: %{title: "{{query_string}}"}}}},
  method: :post
}

validate(body)

Validates a potentially expensive query without executing it.

validate(body, index)

validate(body, index, type)