Elastex v0.1.2 Elastex.Document

Follows the Elasticsearch Document API.

Types

body :: map | struct | nil

int_or_string :: non_neg_integer | String.t

Functions

bulk(builders)

Specs

bulk([Elastex.Builder.t]) :: Elastex.Builder.t

Performs index, create, delete, or update operations in a single call.

Examples

iex> bulk_requests = [
...> Elastex.Document.delete("twitter", "tweet", 5),
...> Elastex.Document.index(%{hello: "world"}, "twitter", "tweet")
...> ]
iex> Elastex.Document.bulk(bulk_requests)
%Elastex.Builder {
  url:    "_bulk",
  body:   "{\"delete\":{\"_type\":\"tweet\",\"_index\":\"twitter\",\"_id\":5}}\n{\"index\":{\"_type\":\"tweet\",\"_index\":\"twitter\"}}\n{\"hello\":\"world\"}\n",
  method: :post,
  action: :document_bulk
}

delete(index, type, id)

Specs

delete(String.t, String.t, int_or_string) :: Elastex.Builder.t

Deletes a document from the index based on its id.

Examples

iex> Elastex.Document.delete("twitter", "tweet", 5)
%Elastex.Builder {
  url:    "twitter/tweet/5",
  method: :delete,
  action: :document_delete,
  index:  "twitter",
  type:   "tweet",
  id:     5
}

exists(index, type, id)

Specs

exists(String.t, String.t, int_or_string) :: Elastex.Builder.t

Checks whether a document exists.

Examples

iex> Elastex.Document.exists("twitter", "tweet", 5)
%Elastex.Builder {
  url:    "twitter/tweet/5",
  method: :head
}

get(index, type, id)

Specs

get(String.t, String.t, int_or_string) :: Elastex.Builder.t

Gets a document from the index based on its id.

Examples

iex> Elastex.Document.get("twitter", "tweet", 5)
%Elastex.Builder {
  url:    "twitter/tweet/5",
  method: :get
}

index(body, index, type)

Specs

index(body, String.t, String.t) :: Elastex.Builder.t

Adds a document in a specific index.

Examples

iex> Elastex.Document.index(%{hello: "world"}, "twitter", "tweet")
%Elastex.Builder {
  url:    "twitter/tweet",
  body:   %{hello: "world"},
  method: :post,
  action: :document_index,
  index:  "twitter",
  type:   "tweet"
}

index(body, index, type, id)

Specs

index(body, String.t, String.t, int_or_string) :: Elastex.Builder.t

Adds a document in a specific index with id.

Examples

iex> Elastex.Document.index(%{hello: "world"}, "twitter", "tweet", 5)
%Elastex.Builder {
  url:    "twitter/tweet/5",
  body:   %{hello: "world"},
  method: :put,
  action: :document_index,
  index:  "twitter",
  type:   "tweet",
  id:     5
}

mget(body)

Specs

mget(body) :: Elastex.Builder.t

Gets multiple documents based on docs array information.

Examples

iex> docs = %{docs: [%{_index: "website", _type: "blog", _id: 2}]}
iex> Elastex.Document.mget(docs)
%Elastex.Builder {
  url:    "_mget",
  body:   %{docs: [%{_index: "website", _type: "blog", _id: 2}]},
  method: :get
}

mget(body, index)

Specs

mget(body, String.t) :: Elastex.Builder.t

Gets multiple documents based on index and docs array information.

Examples

iex> docs = %{docs: [%{_type: "blog", _id: 2}]}
iex> Elastex.Document.mget(docs, "website")
%Elastex.Builder {
  url:    "website/_mget",
  body:   %{docs: [%{_type: "blog", _id: 2}]},
  method: :get,
}

mget(body, index, type)

Specs

mget(body, String.t, String.t) :: Elastex.Builder.t

Gets multiple documents based on index, type and docs array information.

Examples

iex> docs = %{docs: [%{ids: ["1", "2"]}]}
iex> Elastex.Document.mget(docs, "website", "blog")
%Elastex.Builder {
  url:    "website/blog/_mget",
  body:   %{docs: [%{ids: ["1", "2"]}]},
  method: :get
}

mterm_vectors(body)

Specs

mterm_vectors(body) :: Elastex.Builder.t

Gets multiple term vectors at once using body options only.

Examples

iex> body = %{doc: [%{_index: "twitter", _type: "tweet", _id: 2}]}
iex> Elastex.Document.mterm_vectors(body)
%Elastex.Builder {
  url:    "_mtermvectors",
  body:   %{doc: [%{_index: "twitter", _type: "tweet", _id: 2}]},
  method: :get
}

mterm_vectors(body, index)

Specs

mterm_vectors(body, String.t) :: Elastex.Builder.t

Gets multiple term vectors at once using body options and an index.

Examples

iex> body = %{doc: [%{_index: "twitter", _type: "tweet", _id: 2}]}
iex> Elastex.Document.mterm_vectors(body, "twitter")
%Elastex.Builder {
  url:    "twitter/_mtermvectors",
  body:   %{doc: [%{_index: "twitter", _type: "tweet", _id: 2}]},
  method: :get
}

mterm_vectors(body, index, type)

Specs

mterm_vectors(body, String.t, String.t) :: Elastex.Builder.t

Gets multiple term vectors at once using body options, an index and type.

Examples

iex> body = %{doc: [%{_index: "twitter", _type: "tweet", _id: 2}]}
iex> Elastex.Document.mterm_vectors(body, "twitter", "tweet")
%Elastex.Builder {
  url:    "twitter/tweet/_mtermvectors",
  body:   %{doc: [%{_index: "twitter", _type: "tweet", _id: 2}]},
  method: :get
}

term_vectors(index, type, id)

Specs

term_vectors(String.t, String.t, int_or_string) :: Elastex.Builder.t

Returns information and statistics on terms in the fields of a particular document.

Examples

iex> Elastex.Document.term_vectors("twitter", "tweet", "1")
%Elastex.Builder {
  url:    "twitter/tweet/1/_termvectors",
  method: :get
}

term_vectors(body, index, type, id)

Specs

term_vectors(body, String.t, String.t, int_or_string) :: Elastex.Builder.t

Returns information and statistics on terms in the fields of a particular document using body options.

Examples

iex> body = %{fields: ["text"], "positions": true}
iex> Elastex.Document.term_vectors(body, "twitter", "tweet", "1")
%Elastex.Builder {
  url:    "twitter/tweet/1/_termvectors",
  body:   %{fields: ["text"], "positions": true},
  method: :get
}

update(body, index, type, id)

Specs

update(body, String.t, String.t, int_or_string) :: Elastex.Builder.t

Updates a document based on its id.

Examples

iex> Elastex.Document.update(%{doc: %{name: "new_name"}}, "twitter", "tweet", 5)
%Elastex.Builder {
  url:    "twitter/tweet/5/_update",
  body:   %{doc: %{name: "new_name"}},
  method: :post,
  action: :document_update,
  index:  "twitter",
  type:   "tweet",
  id:     5
}