Elastex v0.1.2 Elastex.Index

Functions

add_alias(index, name)

Adds a single alias

add_template(body, template)

Adds a template.

aliases(body)

Aliases an index with a name.

analyze(body)

Performs the analysis process on text and returns the token breakdown.

analyze(body, index)

clear_cache()

Clears all or specific cache associated with an index.

clear_cache(index)

close(index)

Closes an index.

create(index)

Instantiates an index with or without a body.

create(body, index)

delete(index)

Deletes an existing index.

delete_alias(index, name)

Deletes an existing alias

delete_template(template)

Deletes a template.

delete_warmer(index, warmer)

Deletes warmer

exists(index)

Checks if index exists.

flush()

Flushes one or more indices.

flush(index)

force_merge()

Merges one or more indices

force_merge(index)

get(index)

Retrieves information about one or more indices.

get_alias(index, name)

Retrieves existing alias

get_mapping(index)

Retrieves mapping definitions for an index or index/type.

get_mapping(index, type)

get_settings(index)

Retrieves settings of index/indices.

get_template(template)

Retrieves a template

get_warmer(index)

Retrieves warmer

get_warmer(index, warmer)

open(index)

Opens an index

put_mapping(body, index)

Adds a new type to an exisiting index or new fields to an exisiting type.

put_mapping(body, index, type)

put_warmer(body, warmer)

Adds warmer to a specific index.

put_warmer(body, index, warmer)

put_warmer(body, index, type, warmer)

recovery()

Provides insight into on-going index shard recoveries.

recovery(index)

refresh()

Explicitly refreshes one or more index.

refresh(index)

segments()

Gets low level segment information that a Lucene index (shard-level) is built with.

segments(index)

shard_stores()

Provides store information for shard copies of indices.

shard_stores(index)

stats()

Provides statistics on different operations happening on an index.

stats(index)

synced_flush()

Initiates a synched flush manually.

synced_flush(index)

template_exists(template)

Checks the existence of a template.

type_exists(index, type)

Checks if type exist in index.

update_settings(body)

Updates index level settings.

update_settings(body, index)

upgrade()

Upgrades one or more indices

upgrade(index)