Elastex v0.1.2 Elastex.Search

Functions

count(index, type)

Executes a query to get the number of matches for that query.

count(body, index, type)

explain(body, index, type, id)

Computes a score explanation for a query and a specific document.

query()

Executes a search query.

query(body)

query(body, index)

query(body, index, type)

shards(index)

Returns indices and shards that a search request would be executed against.

shards(index, type)

suggest(body)

Suggests similar looking terms based on a provided text.

template(body)

Allows use of mustache templating language to pre-render search requests.

validate(body)

Validates a potentially expensive query without executing it.

validate(body, index)

validate(body, index, type)